@@ -8,9 +8,7 @@ use Chubbyphp\Validation\ValidatorLogicException;
 
 final class MethodAccessor implements AccessorInterface
 {
-    public function __construct(private string $property)
-    {
-    }
+    public function __construct(private string $property) {}
 
     /**
      * @return mixed
