@@ -4,6 +4,4 @@ declare(strict_types=1);
 
 namespace Chubbyphp\Validation\Constraint\Symfony;
 
-final class NotImplementedException extends \RuntimeException
-{
-}
+final class NotImplementedException extends \RuntimeException {}
